ID:5025 Fix sidebar sometimes not animating
SidebarPlugin overwrites some styles and therefore interferes with the
animation.
The class in question is sidebar-overlay. However we do want some other
styles applied to that class, so we can not just rename it.
Adding an additional class to be stronger than SidebarPlugin. However it
would be better to migrate all required styles here.