ID: 5085 Added token based validation during vue instantiation
id: 5085
Henceforth each plugin, e.g. SearchGridPlugin, which depends on
Vue-Components has to register itself before in the backend.
Each plugin receives a one-time nonce which is valid for one session.
The receive token has to be store in the 'data-vue-client-token'
attribute on the vue-conatiner.
Thus were now able to prevent XSS Vue injection.